Fix crash caused by invalid construction ghost shader name (#336)
* Fix construction ghost shader name

Construction ghosts were using a shader called `selection_outline_unshader`, which is not a valid shader name. This caused a crash when constructing things. This fixes that problem by updating them to use a valid shader `unshaded`.

I haven't followed the changes to shaders, but it's likely that this was caused by the shader name being changed and not updated in the construction ghost prototype.

* Remove specific outline shader for construction ghost prototypes

No longer needed.