Syndie Axe description clarity (#18004)
* Update fireaxe.yml

The syndie 'fire axe' already has a lot of drawbacks and is relatively hard to use stealthily, but on the rare occasions you are say an Atmo tech or engi, it should be vaguely possible. This change makes the sprite and name the same as the regular fireaxe, but lets the examine description remain different. 

Blame the sweaty nerds who memorize pixels for making this necessary.

* Clearer syndie fireaxe description

Makes the description detail that the axes are visually distinct.

* Fireaxe Update uplink-catalog.ftl